requirements.txt
alembic==1.13.0
annotated-types==0.6.0
anyio==3.7.1
async-timeout==4.0.3
asyncpg==0.29.0
bcrypt==4.1.2
certifi==2023.11.17
cffi==1.16.0
click==8.1.7
colorama==0.4.6
cryptography==41.0.7
ecdsa==0.18.0
fastapi==0.105.0
greenlet==3.0.2
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
idna==3.6
iniconfig==2.0.0
Mako==1.3.0
MarkupSafe==2.1.3
orjson==3.9.10
packaging==23.2
passlib==1.7.4
pluggy==1.3.0
pyasn1==0.5.1
pycparser==2.21
pydantic==2.5.2
pydantic-settings==2.1.0
pydantic_core==2.14.5
pytest==7.4.3
pytest-asyncio==0.23.3
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.23
starlette==0.27.0
typing_extensions==4.9.0
uvicorn==0.24.0.post1
